moves TYPE to Input component
